Use flock to avoid race when starting multiple vms
Fixes #7

The issue here is that, because we always download the box to a
name derived from the image, one process could race with another
and attempt to download to the same spot. This commit fixes this,
as well as other potential issues with half-extracted files.

The logic goes like this:

1. First acquire a lock of the 'box' path. This is derived from the
image name. If the lock can't be acquired, block until it can. This
can only hang if the download/extraction process of some other
process hangs. In the future, maybe we should put a max time on that.

2. Check that the final file doesn't already exist. If it does,
some other process did the download while we were blocked. So we're
done.

3. Otherwise, download the box from vagranthub to the locked path.
Because we hold an exclusive lock, no other process can interfere
with this process

4. Extract the 'box.img' file to a temporary image path. This is
essentially a '.part' file. The idea here is that if the process
is killed part way through the extraction, we will correctly ignore
it next time a vm starts with that image (note that we _will_ leave
the '.box' and '.part' files lying around)

5. Once the extraction is complete, atomically move the '.part' file
to the final name. After this, the lock is effectively no longer
held.