Add QMP client and clean shutdown
Fixes #43, #40

The logic here is that we establish a QMP connection and send an
ACPI shutdown request. We then wait a bit for the guest to shutdown
cleanly. If it doesn't, we SIGTERM QEMU and exit.

Return codes are handled in the following way:

If QEMU exits prior to SSH, we return with its exit code. Once the
SSH connection closes, we will always return the SSH exit code,
regardless of whether QEMU exits cleanly. This is mainly so the
`-shutdown-timeout=0` case has the effect of preserving the current
behavior. i.e., QEMU is sent SIGTERM (and therefore has a non-zero
exit code) but we get the SSH code which may be zero.