Fix bug where QEMU would exit on SIGINT
This was cause by transient always passing '-qmp', even when
it was not used (i.e., any time we aren't in an ssh console).
Apparently passing 'qmp' causing QEMU to _always_ exit on SIGINT,
event when in '-nographic' mode. We can avoid this by just not
doing the QMP connection unless we are in ssh mode