Make -name flag optional
Fixes #51

This change requires that we have a 'post-run' step when we can
perform operations like this (and -copy-out-after, for example). To
support this, some changes were made to the `run` api. It now
returns None if the underlying processes don't error. In the event
that the SSH connection or QEMU error, it raises TransientProcessError
which contains the returncode.