Add the concept of an ImageSpec
Fixes #53

The argument passed to '-image' is now an ImageSpec. This consists of
a _name_ of the image in the backend, a _protocol_ that is used to
retrieve the image if it does not exist, and a _source_ which is used
with the protocol to do the retrieval (if needed). For example:

    transient run -image myimage,vagrant=centos/7:2004.01 -ssh

Will create a new backend image 'myimage' using the vagrant protocol
to download a CentOS 7 box. Note that if no protocol/source is
provided, 'vagrant' is assumed. So:

    transient run -image generic/alpine38:v3.0.2 -ssh

Is equivalent to:

    transient run -image generic/alpine38:v3.0.2,vagrant=generic/alpine38:v3.0.2 -ssh

This commit also adds a 'frontend' protocol. This will copy a frontend
image to the backend. For example:

   transient run -name example -image newimage,frontend=some-vm@generic/alpine38:v3.0.2 -ssh

This command will copy the per-VM frontend image for the VM named 'some-vm'
that is associated with the backend 'generic/alpine38:v3.0.2' to the backend
under the name 'newimage' (and then start a VM named 'example'). After
running the above command, you can use the new backend image as usual:

    transient run -image newimage -ssh