Add a transient-kernel to use when host is not available
Fixes #56

The issue we're fixing here is that under ubuntu, the host
kernel is not readable. libguestfs needs a kernel to boot
the virtual machine it starts, and it uses the host one
by default.

There are a few ways we would work around this. We could
just instruct the user to chmod their kernel, as there is
no actual security reason to limit reads of the kernel to
root only. However, this approach won't survive a kernel
update. Therefore, we just ship our own kernel and point
libguestfs to that one.

This approach has the added benefit of making this work
under docker, which also obviously doesn't have a kernel
available. Note, however, that due to a libguestfs bug,
there needs to be a /lib/modules directory in the container.
It can be empty, it just needs to exist.