Remove qemu output proxy
We now just pass '-serial stdio -display none' to qemu
and connect stdout. This simplifies the code, and allows
the user to see output during the shutdown. I initially
didn't do this as I was worried about duplicate output.
However, I don't think this is a real issue in practice.