Add support for 'transient build'
Fixes #64

This commit introduces a 'Imagefile' format that allows building
qcow2 files with a syntax similar to a Dockerfile. For example,
given a ubuntu root filesystem 'ubuntu.tar.gz', a bootable
ubuntu qcow2 file could be made with just:

    FROM scratch
    DISK 2gb MBR
    PARTITION 1 FORMAT ext2 MOUNT /

    ADD ubuntu.tar.gz /
    RUN apt-get update
    RUN DEBIAN_FRONTEND=noninteractive apt-get install -y \
           linux-image-generic grub2 init systemd
    RUN passwd -d root
    RUN sed -i 's/GRUB_CMDLINE_LINUX_DEFAULT=".*"/GRUB_CMDLINE_LINUX_DEFAULT="console=ttyS0"/' \
          /etc/default/grub
    RUN echo "/dev/sda / ext2 rw 0 0" >> /etc/fstab
    RUN grub-install /dev/sda
    RUN grub-mkconfig -o /boot/grub/grub.cfg

This is achieved by booting a minimal kernel and initramfs (made
with Buildroot). This 'builder' VM is passed the new disk and
constructs the desired partitions. Typically a root filesystem
image is then passed in to the builder via SCP. Once this
is done, subsequent 'RUN' commands can be executed which
will run a shell command in a chroot on the new disk.

In the future, the FROM command will allow images to be based
on other existing images. For now, however, only 'FROM scratch'
is supported.