Move transient data from environment to tempfile
Finishes fixing #102

The earlier "Allow QEMU to specify random SSH port" mainly resolves
the race in port allocation by allowing the kernel to pick the port.
This works well, but creates an issue with the `transient ssh`
subcommand. Previously we stored the transient process data such
as the VM name and SSH port in the environment of the QEMU process.
But now that we allow the kernel to pick the port, we don't know
what the value will be until after QEMU is started. This, naturally,
means that we can't store that information in the QEMU environment.

So, we need a new way of storing transient data. Ideally, one that
cannot be outdated, does not require complex locking, and is cleaned
up automatically, event if transient is SIGKILL'd. This patch meets
these requirements:

The idea is simple, we create a tempfile in the python process. We
observe the file descriptor of this tempfile and store _that number_
in the environment of the QEMU process. As we spawn QEMU, we pass
in the file descriptor as well. Then, after QEMU has started and we
have established a QMP connection and retrieved the port that was
bound for SSH forwarding, we can write out the data to the tempfile
(from the python process).

Now, when another transient process is looking through the PIDs it
has access to, it observes the sentinel value and reads the file
descriptor. It can then open /proc/<pid>/fd/<fd from environ> and
get a handle to the tempfile from the python process. The second
transient process can then read the data from the tempfile and
gather any necessary information (e.g., the VM name or SSH port).

Once the transient processes exit (or are killed) the tempfile
is automatically cleaned up. There is no database and no locking
required.