Add utils.read_until and simplify SSHFS logic
This is ultimately necessary because a guest can have errors from
their .bashrc that result in other output to stderr which breaks the
previous approach. Now we just read until the desired bytes are seen,
with a timeout.

In order to make the SSHFS code a bit more maintainable, this also
removes one of the 'stages' of verifying the SSHFS setup. The
multi-stage approach was primarily an optimization for cases where a
user has lots of shared folders, but I don't think that is
sufficiently likely for the maintenance burden.