Fix race when connecting to existing VM with shared folders
The `transient ssh` command waits for the '__TRANSIENT_DATA_FD'
to contain valid json before connecting. Because that fd is
filled out after QEMU has started but _before_ shared folders
have been mounted, it was possible for a `transient ssh` command
to connect to a guest that did not yet have the shared folders
mounted. This can be surprising to users, so wait to fill the
fd until _after_ we have (probably) finished mounting the shared
folders.

In general it is impossible to totally avoid this race this way,
but it should significantly reduce the window.