Add -wait flag to 'transient ssh' and change default behavior
The `transient ssh` command now fails immediately if a VM with the
requested name is not available. To wait until such a VM exists
(to a max timeout of `-ssh-timeout` seconds), the user must now
specify the `-wait` flag. This is useful most often in scripts
where a VM is being started and backgrounded, and a user is
connecting to that VM right after, when it has probably not had
time to fully start yet.