Transition to docker-like ui
Fixes #138

Previously, in order to be consistent with QEMU's commandline
format, transient used single '-' flags like '-ssh', and attempted
to avoid any notion of having explicit VM state. Instead state was
stored in the disk images being used and the command used to
run the initial VM had to be run again to 'restart' the VM.

Now transient uses an interface designed to duplicate docker. While
transient still avoids maintaining extra state in general, but it
_does_ not allow the option of using a 'create' subcommand to
make a VM that persists on disk. This VM can be started and stopped
without needing to know what original command started it.

Additionally, some of the more complex features of transient have
been simplified. For example, there is no longer a 'frontend'
image protocol. In order to build a backend image from a given VM
disk, simply use the `transient commit` command.