Fix a bug where 'make test' would fail
This appears to have been caused by a bug in behave, where setting
a tag causes a change in PWD. Since
51b85af87b3b789277ece05eafa3c36b93bc43bd we have used this variable
as part of the test paths, but this bug causes the effective path
to be incorrect in the 'make test' case, where no tag is set. This
is not seen in CI as all the tests are run separately with targets
that _do_ set a tag.