Fix potential crash while using the frontend image protocl
Currently, transient used the 'acutal_size' field of the existing
frontend image object to get the image size when showing the copy
progress bar. This is populated using the 'actual-size' output
of qemu-img. However, this appears to round _down_ to some kind
of block boundry. This causes an incorrect size calculation and
a crash when the actual number of bytes copied exceeds the max
value.

This commit switches to directly determining the file size.