Allow compatible versions of dependencies that are 1.0+
Fixes #112

This allows major-version compatible version of all dependencies.
It is not possible to do the same for some packages, because they
are <1.0 and therefore (by semantic versioning), they may change
their interface in any release.

We should check these dependencies periodically to see if any of
them have become stable.