Update to Poetry from setup.py
The setup.py-style builds are deprecated. PEP 517 introduced the pyproject.toml
way of packing Python projects. This commit switches to this new mechanism and
uses the Poetry build system. The advantage of this system is that a
Rust-inspired lock file is created to lock down exact dependencies and all but
guarantees that future users will be able to build and use the project.

Further, this commit fixes up CI to use Poetry.

Support for Python 3.6 is dropped since it is no longer supported and the
gherkin-formatter does not want to play nice with 3.6.

Testing against Python 3.10 is added.

Other file changes are reformatting due to new formatter versions.